The floor was never the problem: the pad, and the doors
Session 8's three items, all in.  The first was a measurement rather than a
mystery.

The ground gap.  Session 8's per-set table - 348 units in PRI, 155 in CA, and a
414-unit character buried to the hips - was an artefact of its own estimator.
That estimator took the modal reconstructed height over a collision triangle's
whole plan footprint, which holds the wall standing on the triangle's edge and
the crate in the middle of it; measured on the same triangle from two cameras it
disagrees with itself by a median of 300 units, more than the gap it reports.
Nothing is below the floor, so measure the lowest flat surface a camera sees in
a patch of plan instead: over 43 sets that is median -5, sd 61, with no per-set
structure, PRI falling from 348 to 45.

Then measured a fourth time by the engine, with no backprojection at all: raise
a model until the backdrop's Z-buffer stops hiding any of it.  In PRI the
69-unit bottle at collision height 1 keeps 2 of 23 pixels and is clear by y=60.
Run at several places in one set it does not agree with itself either - CA gives
20, 50, 110, 190 - and that disagreement is the answer.  The gap is local relief
in the art of a few tens of units, six or seven pixels on a 320x200 screen.

The surviving .MAP files say why there was never a correction to find.  They are
Matthew Jesson's map editor's own format, and a set is a plan traced over a
bitmap with one shared vertex list: world = (pixel - origin) * DISTANCE *
ZOOMDIV / (ZOOMMUL * |v0-v1|), which for DUN1 is 39.49 units a pixel and lands
182 of the disc's 217 collision vertices within 20 units.  There is no vertical
calibration in the format anywhere.  Every ground height in the game is an
integer somebody typed into a HEIGHT field by eye.  The maps also settle what
10.5 left open: an event is a line segment there, and the exporter reduces each
one to a handful of circles on it - 64 segments to 77 circles in DUN1, against
segments up to 9,054 units long and radii of 13 to 57.

The player, driven.  src/game/control.c is AICTRL.GAS's two loops.  PlayerControl
forces JOY_UP back on for three frames after release, in a branch delay slot, so
a brief release does not stop the walk - and a press inside that window is the
double tap, detected by the pad handler rather than the animation table.
ActionCode walks a logic table of tuples headed by joypad masks, first match
wins, and citStance & 7 picks between eight entries, so what a button does
depends on what the character is already doing.  Turning is 4 * 20 / framerate
steps a frame unless the stance carries FSATurn - and Quentin's turn-on-the-spot
animation turns 4.1 a frame, so the two rates are the same rate.  Which
animation is which is read off the root motion, not guessed.  The table itself
is ours: the real ones are five CD files DATA.INC names, at block offsets of the
data type VIDSTUFF.INC calls CHARDATA and annotates "jakes joypad logics", and
where the retail disc put them is not known.

The doorways.  10.3 runs the other way round from how it was written: the init
entry's id is the view you are LEAVING.  Not one of the disc's 153 entries is
keyed on its own set's group, and for all 123 with a real id the set that owns
that id has a SCENE event, fired from that very view, cutting into the arriving
set.  The flags word decodes too - facing * 256 + character, for all 153 - which
the .MAP says in words as two START blocks, QUENTIN and RAMIREZ, sharing one
FROM, each with an ORIENTATION in degrees.  hlview --check-doors runs it over
the disc: 153 of 153 entries decode and stand on their own mesh, and all 129
cross-set cuts find an arrival, none off the far mesh.

That check found a bug that had nothing to do with doorways.  A triangle's
vertex index was being read into a uint8_t; the disc stores a word, and CNY01
has 348 vertices, D1 300 and NEOSW 293.  Three meshes were silently folded, and
--check-mesh could not see it because the search and the scan it is compared
against were both reading the same folded mesh.

And SHEET.S closes 12.7's first open item for free: the run at cshFileOff is a
table of (entry position, data type, block offset), which is what links a
character sheet to its blobs on the CD.  Quentin's thirty animations are three
loads of 14, 14 and 2, and the sword and gun banks live on the weapon's sheet -
session 8's "four banks, 114 in all", explained.

+++ src/game/set.c
@@ -99,7 +99,7 @@ int set_load(Set *s, const char *track3, int index)
     for (int i = 0; i < s->ntris; i++) {
         const uint8_t *p = d + coll + troff + (size_t)i * 14;
         s->tri[i].height = be16s(p);
-        for (int k = 0; k < 3; k++) s->tri[i].vert[k] = (uint8_t)be16(p + 2 + k * 2);
+        for (int k = 0; k < 3; k++) s->tri[i].vert[k] = be16(p + 2 + k * 2);
         for (int k = 0; k < 3; k++) s->tri[i].adj[k]  = be16s(p + 8 + k * 2);
     }
